+++ server/json/user1-institutions.json
@@ -3,6 +3,7 @@
     {
       "name": "Bank 0",
       "id": "bank0id",
+      "period": "2017",
       "status": {
         "code": 0,
         "message": ""
@@ -12,6 +13,7 @@
     {
       "name": "Bank 1",
       "id": "bank1id",
+      "period": "2017",
       "status": {
         "code": -1,
         "message": "Submission cancelled by user"
@@ -21,6 +23,7 @@
     {
       "name": "Bank 2",
       "id": "bank2id",
+      "period": "2017",
       "status": {
         "code": 3,
         "message": ""
@@ -30,6 +33,7 @@
     {
       "name": "Bank 3",
       "id": "bank3id",
+      "period": "2017",
       "status": {
         "code": 6,
         "message": ""
@@ -39,6 +43,7 @@
     {
       "name": "Bank 4",
       "id": "bank4id",
+      "period": "2017",
       "status": {
         "code": 9,
         "message": ""
@@ -48,6 +53,7 @@
     {
       "name": "Bank 5",
       "id": "bank5id",
+      "period": "2017",
       "status": {
         "code": 11,
         "message": ""
