+++ server/json/user2-institutions.json
@@ -3,6 +3,7 @@
     {
       "name": "Another procrastinator",
       "id": "123",
+      "period": "2017",
       "status": {
         "code": 0,
         "message": ""
@@ -12,6 +13,7 @@
     {
       "name": "Getting coffee",
       "id": "124",
+      "period": "2017",
       "status": {
         "code": 4,
         "message": ""
@@ -21,6 +23,7 @@
     {
       "name": "Wacky data",
       "id": "125", 
+      "period": "2017",
       "status": {
         "code": 6,
         "message": ""
@@ -30,6 +33,7 @@
     {
       "name": "Responsible fi",
       "id": "126",  
+      "period": "2017",
       "status": {
         "code": 11,
         "message": ""
