--- server/json/year.json
@@ -1,5 +1,0 @@
-{
-  "year": "2017",
-  "spec": "speclink.gov",
-  "filingDate": "20180301"
-}
